Atualização LOPAL Ex 2.ino

+++ Exercícios de Lógica/LOPAL Ex 2.ino
@@ -4,10 +4,10 @@
 //b. Essa idade convertida em semana
 
 //Variáveis 
-long anoNascimento = 0;
-long anoAtual = 0;
-long idadeAtual = 0;
-long idadeSemana = 0;
+int anoNascimento = 0;
+int anoAtual = 0;
+int idadeAtual = 0;
+int idadeSemana = 0;
 
 void setup()
 {
